requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-filter==24.3
django-phonenumber-field==8.0.0
django-resized==1.0.3
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
phonenumberslite==8.13.53
pillow==11.1.0
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1